Move parallel pull implementation to separate class
