+++ cli/src/main/java/com/box/l10n/mojito/cli/command/CommandHelper.java
@@ -289,8 +289,7 @@ public class CommandHelper {
   public void waitForPollableTaskSilencedOutput(Long pollableId) throws CommandException {
 
     try {
-      pollableTaskClient.waitForPollableTask(
-          pollableId, PollableTaskClient.NO_TIMEOUT, new CommandWaitForPollableTaskListener(true));
+      pollableTaskClient.waitForPollableTask(pollableId, PollableTaskClient.NO_TIMEOUT, null);
     } catch (PollableTaskException e) {
       throw new CommandException(e.getMessage(), e.getCause());
     }
