+++ cli/src/main/java/com/box/l10n/mojito/cli/command/CommandWaitForPollableTaskListener.java
@@ -29,28 +29,16 @@ public class CommandWaitForPollableTaskListener implements WaitForPollableTaskLi
    */
   int numberAfterPollCallback = 0;
 
-  boolean silent;
-
-  CommandWaitForPollableTaskListener() {
-    this.silent = false;
-  }
-
-  CommandWaitForPollableTaskListener(boolean silent) {
-    this.silent = silent;
-  }
-
   @Override
   public void afterPoll(PollableTask pollableTask) {
 
-    if (!silent) {
-      if (numberAfterPollCallback++ > 0) {
-        consoleWriter.erasePreviouslyPrintedLines();
-      }
-
-      printPollableTaskMessages(pollableTask, 0);
-
-      consoleWriter.print();
+    if (numberAfterPollCallback++ > 0) {
+      consoleWriter.erasePreviouslyPrintedLines();
     }
+
+    printPollableTaskMessages(pollableTask, 0);
+
+    consoleWriter.print();
   }
 
   /**
